Renamed i386 folder to x86
